+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 
 - Autocomplete: Open flyout on click after an item was selected when input is already focused
 - Container: Updated to content-box to match the passed width value
+- Overlay: containerEl locks scroll only for itself and not the body
+- ScrollArea: Fixed nested scroll area hover
+- Accordion: fixed animation in Preact